+++ src/mqtt-broker/src/server/tcp/handler.rs
@@ -1,3 +1,17 @@
+// Copyright 2023 RobustMQ Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 use crate::{
     handler::command::Command,
     server::{
@@ -5,11 +19,8 @@ use crate::{
         packet::{RequestPackage, ResponsePackage},
     },
 };
-
 use common_base::error::mqtt_broker::MQTTBrokerError;
-
 use log::{debug, error, info};
-
 use std::{collections::HashMap, sync::Arc};
 use storage_adapter::storage::StorageAdapter;
 use tokio::sync::broadcast;
@@ -56,7 +67,6 @@ pub(crate) async fn handler_process<S>(
                 },
                 val = request_queue_rx.recv()=>{
                     if let Some(packet) = val{
-                        println!("{:?}",packet);
                         // Try to deliver the request packet to the child handler until it is delivered successfully.
                         // Because some request queues may be full or abnormal, the request packets can be delivered to other child handlers.
                         loop{
